Restricted User ID inputs and fixed delete team sql
User inputs now auto-populate with the user_id of the session and cannot be changed.
Delete team sql statements only found team_id, but 2 users can have a team with the same id, so I also grabbed the user_id from the session.

+++ deleteTeam.php
@@ -1,17 +1,19 @@
 <?php
+session_start();
 include 'pokeHeader.php';
 
 echo '<html>';
 echo '<body>';
 
 $team_id = $_GET['team_id'];
+$user_id = $_SESSION["user_id"];
 
 // Delete team members first
-$sql = "DELETE FROM Team_Members WHERE team_id = '$team_id'";
+$sql = "DELETE FROM Team_Members WHERE team_id = '$team_id' AND user_id = '$user_id'";
 $link->query($sql);
 
 // Delete the team
-$sql = "DELETE FROM Teams WHERE team_id = '$team_id'";
+$sql = "DELETE FROM Teams WHERE team_id = '$team_id' AND user_id = '$user_id'";
 if ($link->query($sql) === TRUE) {
     echo "Team deleted successfully";
 } else {
